Remove unused and refactor to use useFetchTransactionStatus
