+++ packages/widget-v2/src/pages/SwapExecutionPage/SwapExecutionPage.tsx
@@ -2,11 +2,11 @@ import { Column } from "@/components/Layout";
 import { MainButton } from "@/components/MainButton";
 import { SwapPageFooter } from "@/pages/SwapPage/SwapPageFooter";
 import { SwapPageHeader } from "@/pages/SwapPage/SwapPageHeader";
-import { useEffect, useMemo, useState } from "react";
+import { useMemo, useState } from "react";
 import { ICONS } from "@/icons";
 import { SetAddressModal } from "@/modals/SetAddressModal/SetAddressModal";
 import { useTheme } from "styled-components";
-import { useAtom, useAtomValue, useSetAtom } from "jotai";
+import { useAtomValue, useSetAtom } from "jotai";
 import { SwapExecutionPageRouteSimple } from "./SwapExecutionPageRouteSimple";
 import { SwapExecutionPageRouteDetailed } from "./SwapExecutionPageRouteDetailed";
 
@@ -14,20 +14,16 @@ import { useModal } from "@/components/Modal";
 import { currentPageAtom, Routes } from "@/state/router";
 import {
   ClientOperation,
-  ClientTransferEvent,
   getClientOperations,
-  getOperationToTransferEventsMap,
-  getTransferEventsFromTxStatusResponse,
 } from "@/utils/clientType";
 import {
   chainAddressesAtom,
-  setOverallStatusAtom,
   skipSubmitSwapExecutionAtom,
-  skipTransactionStatusAtom,
   swapExecutionStateAtom,
 } from "@/state/swapExecutionPage";
 import { useAutoSetAddress } from "@/hooks/useAutoSetAddress";
 import { convertSecondsToMinutesOrHours } from "@/utils/number";
+import { useFetchTransactionStatus } from "./useFetchTransactionStatus";
 import { getSignRequiredChainIds } from "@/utils/operations";
 
 enum SwapExecutionState {
@@ -42,19 +38,17 @@ enum SwapExecutionState {
 export const SwapExecutionPage = () => {
   const theme = useTheme();
   const setCurrentPage = useSetAtom(currentPageAtom);
-  const setOverallStatus = useSetAtom(setOverallStatusAtom);
-  const { route, transactionDetailsArray, overallStatus } = useAtomValue(
-    swapExecutionStateAtom
-  );
+  const {
+    route,
+    overallStatus,
+  } = useAtomValue(swapExecutionStateAtom);
   const chainAddresses = useAtomValue(chainAddressesAtom);
   const { connectRequiredChains } = useAutoSetAddress();
-  const [{ data: transactionStatus }] = useAtom(skipTransactionStatusAtom);
   const [simpleRoute, setSimpleRoute] = useState(true);
   const setManualAddressModal = useModal(SetAddressModal);
 
-  const { mutate, isPending } = useAtomValue(skipSubmitSwapExecutionAtom);
-  const [operationToTransferEventsMap, setOperationToTransferEventsMap] =
-    useState<Record<number, ClientTransferEvent>>({});
+  const { mutate } = useAtomValue(skipSubmitSwapExecutionAtom);
+  const operationToTransferEventsMap = useFetchTransactionStatus();
 
   const clientOperations = useMemo(() => {
     if (!route?.operations) return [] as ClientOperation[];
@@ -65,87 +59,6 @@ export const SwapExecutionPage = () => {
     const signRequiredChains = getSignRequiredChainIds(clientOperations);
     return signRequiredChains;
   }, [clientOperations]);
-
-  const computedSwapStatus = useMemo(() => {
-    const operationTransferEventsArray = Object.values(
-      operationToTransferEventsMap
-    );
-
-    if (operationTransferEventsArray.length === 0) {
-      if (isPending) {
-        setOverallStatus("signing");
-      }
-      return;
-    }
-
-    if (
-      operationTransferEventsArray.every(({ status }) => status === "completed")
-    ) {
-      if (operationTransferEventsArray.length === route?.operations.length) {
-        return "completed";
-      }
-      return "pending";
-    }
-
-    if (
-      operationTransferEventsArray.find(({ status }) => status === "failed")
-    ) {
-      return "failed";
-    }
-    if (
-      operationTransferEventsArray.find(({ status }) => status === "pending")
-    ) {
-      return "pending";
-    }
-    if (
-      operationTransferEventsArray.every(
-        (state) => state.status === "broadcasted"
-      )
-    ) {
-      return "broadcasted";
-    }
-  }, [
-    isPending,
-    operationToTransferEventsMap,
-    route?.operations.length,
-    setOverallStatus,
-  ]);
-
-  useEffect(() => {
-    if (overallStatus === "completed" || overallStatus === "failed") return;
-
-    const transferEvents =
-      getTransferEventsFromTxStatusResponse(transactionStatus);
-    const operationToTransferEventsMap = getOperationToTransferEventsMap(
-      transactionStatus ?? [],
-      clientOperations
-    );
-    const operationTransferEventsArray = Object.values(
-      operationToTransferEventsMap
-    );
-
-    if (transactionDetailsArray.length === 1 && transferEvents.length === 0) {
-      setOperationToTransferEventsMap({
-        0: {
-          status: "pending",
-        } as ClientTransferEvent,
-      });
-    }
-    if (operationTransferEventsArray.length > 0) {
-      setOperationToTransferEventsMap(operationToTransferEventsMap);
-    }
-
-    if (computedSwapStatus) {
-      setOverallStatus(computedSwapStatus);
-    }
-  }, [
-    clientOperations,
-    overallStatus,
-    computedSwapStatus,
-    setOverallStatus,
-    transactionDetailsArray.length,
-    transactionStatus,
-  ]);
 
   const swapExecutionState = useMemo(() => {
     if (!chainAddresses) return;
